My places and My trails added
added pages with user places and trails;
Each place and trail have acces to previe,edit and delete (not all functions ready);
Renamed objectDetails to placeDetails;
Renamed profile_my_places to myPlaces;
Fixed login display;
Fixed trail parse;
Fixed place object;
Fixed profile button.

+++ memo_places_mobile/lib/Objects/trail.dart
@@ -35,8 +35,8 @@ class Trail {
   factory Trail.fromJson(Map<String, dynamic> json) {
     List<dynamic> coordinatesJson = jsonDecode(json['coordinates']);
     List<LatLng> coordinates = coordinatesJson.map((coord) {
-      double lat = coord['lat'];
-      double lng = coord['lng'];
+      double lat = coord['lat'] as double;
+      double lng = coord['lng'] as double;
       return LatLng(lat, lng);
     }).toList();
